requirements.txt
# Core LangChain & LangGraph
langchain==0.1.0
langchain-community==0.0.13
langgraph==0.0.20
langchain-core==0.1.10

# LLM Integration
ollama==0.1.6

# Embeddings & Vector Stores
sentence-transformers==2.3.1
faiss-cpu==1.7.4
chromadb==0.4.22

# Data Processing
pandas==2.2.0
numpy==1.26.3
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
tqdm==4.66.1
requests==2.31.0
beautifulsoup4==4.12.3

# Geospatial (lightweight, no external APIs)
geopy==2.4.1
scikit-learn==1.4.0  # For clustering (DBSCAN/KMeans)

# Date/Time handling
python-dateutil==2.8.2
pytz==2023.3

# CLI & Formatting
rich==13.7.0
typer==0.9.0

# Development
pytest==7.4.4
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Optional: Jupyter for exploration
jupyter==1.0.0
matplotlib==3.8.2